Update to bevy 0.16, clippy

+++ src/builtins/climb.rs
@@ -105,7 +105,7 @@ impl TnuaAction for TnuaBuiltinClimb {
                 TnuaBuiltinClimbState::Climbing { climbing_velocity } => {
                     if matches!(lifecycle_status, TnuaActionLifecycleStatus::NoLongerFed) {
                         *state = TnuaBuiltinClimbState::Coyote(Timer::from_seconds(
-                            self.coyote_time as f32,
+                            self.coyote_time,
                             TimerMode::Once,
                         ));
                         continue;
